+++ flexmeasures/data/schemas/scheduling/__init__.py
@@ -1,4 +1,4 @@
-from marshmallow import Schema, fields, validate
+from marshmallow import Schema, fields, validate, validates_schema, ValidationError
 
 from flexmeasures.data.schemas.sensors import (
     VariableQuantityField,
@@ -29,8 +29,34 @@ class FlexContextSchema(Schema):
         data_key="site-consumption-capacity",
         validate=validate.Range(min=0),
     )
+    # todo: deprecated since flexmeasures==0.23
     consumption_price_sensor = SensorIdField(data_key="consumption-price-sensor")
     production_price_sensor = SensorIdField(data_key="production-price-sensor")
+
+    consumption_price = VariableQuantityField(
+        "/MWh",
+        required=False,
+        data_key="consumption-price",
+        return_magnitude=False,
+    )
+    production_price = VariableQuantityField(
+        "/MWh",
+        required=False,
+        data_key="production-price",
+        return_magnitude=False,
+    )
     inflexible_device_sensors = fields.List(
         SensorIdField(), data_key="inflexible-device-sensors"
     )
+
+    @validates_schema
+    def check_prices(self, data: dict, **kwargs):
+        """Check whether the flex-context contains at most 1 consumption price and at most 1 production price field."""
+        if "consumption_price_sensor" in data and "consumption_price" in data:
+            raise ValidationError(
+                "Must pass either consumption-price or consumption-price-sensor."
+            )
+        if "production_price_sensor" in data and "production_price" in data:
+            raise ValidationError(
+                "Must pass either production-price or production-price-sensor."
+            )