+++ flexmeasures/data/schemas/tests/test_sensor.py
@@ -9,34 +9,81 @@ from marshmallow import ValidationError
 
 
 @pytest.mark.parametrize(
-    "sensor_id, src_quantity, dst_unit, fails",
+    "src_quantity, dst_unit, fails, exp_dst_quantity",
     [
         # deserialize a sensor
-        (1, None, "MWh", False),
-        (1, None, "kWh", False),
-        (1, None, "kW", False),
-        (1, None, "EUR", True),
-        (2, None, "EUR/kWh", False),
-        (2, None, "EUR", True),
+        ({"sensor": 1}, "MWh", False, None),
+        ({"sensor": 1}, "kWh", False, None),
+        ({"sensor": 1}, "kW", False, None),
+        ({"sensor": 1}, "EUR", True, None),
+        ({"sensor": 1}, "/h", False, None),  # convertable to MWh²/h
+        ({"sensor": 2}, "EUR/kWh", False, None),
+        ({"sensor": 2}, "EUR", True, None),
         # deserialize a quantity
-        (None, "1MWh", "MWh", False),
-        (None, "1 MWh", "kWh", False),
-        (None, "1 MWh", "kW", True),
-        (None, "100 EUR/MWh", "EUR/kWh", False),
-        (None, "100 EUR/MWh", "EUR", True),
+        ("1MWh", "MWh", False, "1 MWh"),
+        ("1 MWh", "kWh", False, "1000.0 kWh"),
+        ("1 MWh", "kW", True, None),
+        ("100 EUR/MWh", "EUR/kWh", False, "0.1 EUR/kWh"),
+        ("100 EUR/MWh", "EUR", True, None),
+        ("1 EUR/kWh", "/MWh", False, "1.0 kEUR/MWh"),
+        ("50%", "/MWh", False, "500.0 kWh/MWh"),
+        ("/", "/MWh", True, None),
+        ("/", "MWh", True, None),
+        ("10 batteries", "MWh", True, None),
+        # deserialize a time series specification
+        (
+            [{"start": "2024-08-17T11:00+02", "duration": "PT1H", "value": "2 MWh"}],
+            "kWh",
+            False,
+            "2000.0 kWh",
+        ),
+        (
+            [
+                {
+                    "start": "2024-08-17T11:00+02",
+                    "duration": "PT1H",
+                    "value": "829.4 Wh/kWh",
+                }
+            ],
+            "/MWh",
+            False,
+            "829.4 kWh/MWh",
+        ),
+        (
+            [
+                {
+                    "start": "2024-08-17T11:00+02",
+                    "duration": "PT1H",
+                    "value": "914.7 EUR/kWh",
+                }
+            ],
+            "/MWh",
+            False,
+            "914.7 kEUR/MWh",
+        ),
+        # todo: uncomment after to_preferred gets rid of mEUR
+        # (
+        #     [{"start": "2024-08-17T11:00+02", "duration": "PT1H", "value": "120.8 EUR/MWh"}],
+        #     "/kWh",
+        #     False,
+        #     "0.1208 EUR/kWh",
+        # ),
     ],
 )
 def test_quantity_or_sensor_deserialize(
-    setup_dummy_sensors, sensor_id, src_quantity, dst_unit, fails
+    setup_dummy_sensors, src_quantity, dst_unit, fails, exp_dst_quantity
 ):
 
-    schema = QuantityOrSensor(to_unit=dst_unit)
+    schema = VariableQuantityField(to_unit=dst_unit, return_magnitude=False)
 
     try:
-        if sensor_id is None:
-            schema.deserialize(src_quantity)
-        else:
-            schema.deserialize({"sensor": sensor_id})
+        dst_quantity = schema.deserialize(src_quantity)
+        if isinstance(src_quantity, ur.Quantity):
+            assert dst_quantity == ur.Quantity(exp_dst_quantity)
+            assert str(dst_quantity) == exp_dst_quantity
+        elif isinstance(src_quantity, list):
+            assert dst_quantity[0]["value"] == ur.Quantity(exp_dst_quantity)
+            assert str(dst_quantity[0]["value"]) == exp_dst_quantity
         assert not fails
     except ValidationError as e:
         assert fails, e